backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "timm>=1.0.17",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "scikit-learn>=1.7.0",
    "fastapi[all]>=0.116.1",
    "uvicorn>=0.35.0",
    "pywebview>=5.4",
]
